CBG-5784: move tests for channel cache, feed and changes feed into sub-packages
