+++ rest/changes_api_test.go
@@ -1,0 +1,77 @@
+//  Copyright 2022-Present Couchbase, Inc.
+//
+//  Use of this software is governed by the Business Source License included
+//  in the file licenses/BSL-Couchbase.txt.  As of the Change Date specified
+//  in that file, in accordance with the Business Source License, use of this
+//  software will be governed by the Apache License, Version 2.0, included in
+//  the file licenses/APL2.txt.
+
+package rest
+
+import (
+	"testing"
+	"time"
+
+	"github.com/couchbase/sync_gateway/base"
+	"github.com/couchbase/sync_gateway/testing/assert"
+)
+
+func TestReadChangesOptionsFromJSON(t *testing.T) {
+
+	ctx := base.TestCtx(t)
+	h := &handler{}
+	h.server = NewServerContext(ctx, &StartupConfig{}, false)
+	defer h.server.Close(ctx)
+
+	// Basic case, no heartbeat, no timeout
+	optStr := `{"feed":"longpoll", "since": "123456:78", "limit":123, "style": "all_docs",
+				"include_docs": true, "filter": "Melitta", "channels": "ABC,BBC"}`
+	feed, options, filter, channelsArray, _, _, err := h.readChangesOptionsFromJSON([]byte(optStr))
+	assert.NoError(t, err)
+	assert.Equal(t, "longpoll", feed)
+
+	assert.Equal(t, uint64(78), options.Since.Seq)
+	assert.Equal(t, uint64(123456), options.Since.TriggeredBy)
+	assert.Equal(t, 123, options.Limit)
+	assert.Equal(t, true, options.Conflicts)
+	assert.Equal(t, true, options.IncludeDocs)
+	assert.Equal(t, uint64(kDefaultHeartbeatMS), options.HeartbeatMs)
+	assert.Equal(t, uint64(kDefaultTimeoutMS), options.TimeoutMs)
+
+	assert.Equal(t, "Melitta", filter)
+	assert.Equal(t, []string{"ABC", "BBC"}, channelsArray)
+
+	// Attempt to set heartbeat, timeout to valid values
+	optStr = `{"feed":"longpoll", "since": "1", "heartbeat":30000, "timeout":60000}`
+	_, options, _, _, _, _, err = h.readChangesOptionsFromJSON([]byte(optStr))
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(30000), options.HeartbeatMs)
+	assert.Equal(t, uint64(60000), options.TimeoutMs)
+
+	// Attempt to set valid timeout, no heartbeat
+	optStr = `{"feed":"longpoll", "since": "1", "timeout":2000}`
+	_, options, _, _, _, _, err = h.readChangesOptionsFromJSON([]byte(optStr))
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(2000), options.TimeoutMs)
+
+	// Disable heartbeat, timeout by explicitly setting to zero
+	optStr = `{"feed":"longpoll", "since": "1", "heartbeat":0, "timeout":0}`
+	_, options, _, _, _, _, err = h.readChangesOptionsFromJSON([]byte(optStr))
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(0), options.HeartbeatMs)
+	assert.Equal(t, uint64(0), options.TimeoutMs)
+
+	// Attempt to set heartbeat less than minimum heartbeat, timeout greater than max timeout
+	optStr = `{"feed":"longpoll", "since": "1", "heartbeat":1000, "timeout":1000000}`
+	_, options, _, _, _, _, err = h.readChangesOptionsFromJSON([]byte(optStr))
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(kMinHeartbeatMS), options.HeartbeatMs)
+	assert.Equal(t, uint64(kMaxTimeoutMS), options.TimeoutMs)
+
+	// Set max heartbeat in server context, attempt to set heartbeat greater than max
+	h.server.Config.Replicator.MaxHeartbeat = base.NewConfigDuration(time.Minute)
+	optStr = `{"feed":"longpoll", "since": "1", "heartbeat":90000}`
+	_, options, _, _, _, _, err = h.readChangesOptionsFromJSON([]byte(optStr))
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(60000), options.HeartbeatMs)
+}
